sierrachart: add default studies package
The file SierraChart/Data/UserContributedStudies_64.dll
was not being updated since it's only copied during the
first execution of the launcher script. Therefore it was
possible for the DLL to become outdated.

This change creates a new "study package" which installs the
aforementioned DLL. This study package is always installed.
That way, the DLL is kept up-to-date with the one provided
by upstream.