mkWindowsApp: make WIN_LAYER_HASH precise
Instead of using a flexible hash, use the hash of:

* The Nix store path to Wine.
* WA_API

This will produce Windows layer rebuilds more often, but it avoids
potential problems. For example, with the old method, if the Wine
package is changed and the user already has a Windows layer that was
compatible. Then when the user runs the app it will use the old,
incorrect Windows layer.