mkwindowsapp: share bottles between instances
Some applications behave differently when they detect other instances.
For example, when Notepad++ is already running and a new instance is
executed with a file path, Notepad++ will open a new tab in the original
instance. Then, the new instance terminates.

Up until now, each instance was isolated within its own ephemeral Wine
bottle. Using the Notepad++ example, the new instance would open the
file and remain running.

This commit changes the bottle creation process so that if additional
instances are executed, they share the same bottle. The last instance to
exit cleans up the bottle.

Builds are identified by a hash of the Windows and app layers.
This means that if either input changes, then instances are not
considered instances of the same app, and would therefore use separate
bottles. In short, the same app built with different versions of Nixpkgs
would produce different builds.

This change also introduces two hooks into the derivation function:

- winAppPreRun: This code runs before winAppRun, but only if it hasn't
  been executed already.
- winAppPostRun: This code runs after winAppRun, but only if it hasn't
  been executed already.

The idea behind these hooks is that the 1st instance runs them, but any
other instances do not. It's also expected that winAppRun does not
block.