mkWindowsApp: add registry injection
mkWindowsApp has had a problem loading changes made to the registry
during the layer building process. In short, the changes are saved to
the layer, but for some reason Wine loads a clean registry. I suspect it
has to do with an interaction between Wine and overlayfs. My work-around
has been to apply the needed registry changes right before running the
app.

This change mitigates the problem by "injecting" (aka. copying) the
registry files from the top-most layer of the read-only side of
overlayfs to the mountpoint. This action writes the registry files to
the writeable layer, the overlayfs upper layer, after which Wine is
quite happy to use them. In short, this means the registry files
persisted in the windows layer are made available when building the app
layer, and the registry files persisted in the app layer are made
available to the app at runtime :)