roblox: remove Version directory mapping
My thought process behind adding the Version directory mapping was that
Roblox was continously updating itself because the image was being built
with an old version. But I'm installing the latest version, so that's
not the issue. The mapping duplicates a ton of data, so I'm happy to remove it.

The broken vulkan support (DXVK) suggests that the issue is actually
registry persistence, which should not be resolved through mkWindowsApp
"registry injection."
it.