mkWindowsApp: add optional persistent WINEPREFIX
mkWindowsApp was designed to provide highly-reproducable builds for
Wine-compatible Windows applications; Mimicking the benefits that Nix
provides to Linux applications.

However, there are some applications that don't fit this model; I'm
speaking about you, Roblox!

Applications which automatically update, and for which there's no real
way to disable the automatic updates, an ephemeral WINEPREFIX can be a
problem.

This change adds the ability to preserve the WINEPREFIX that's used at
runtime. Unlike the Windows and App layers which are read-only, the
"Runtime layer" is made read-write.

Keep in mind though, the layer dependency is maintained. Meaning, that
if the Windows or App layers need to be rebuilt, then the Runtime layer
    will be discarded and also rebuilt. Therefore, this option only
    provides short-term persistence.