mkWindowsApp: replace overlayfs with unionfs
Tests confirm that unionfs doesn't suffer from the same copy-up issue I
encountered with overlayfs; Files which exist in the lower layer and
are modified in a higher layer, are not reflecting the changes when
mounted.

Registry injection was my work-around. But that only handles the
registry. The same problem occurs when using `winetricks`, installing
DXVK, etc. In the case of DXVK, I compared the checksum of one of the
Direct3D DLLs and the one in the app layer did not match the one in the
runtime layer. I had to run the DXVK installer during runtime to get the
correct DLLs to show up. And these are just some of the issues I know
of.

The issue may lie in the fact that I didn't use the app layer work
directory when mounting the bottle for runtime. Perhaps I should have,
however Docker actually merges the lower layers into a different
directory when using the overlay2 driver. To me, that suggests it's
unwise to reuse the work directory.

unionfs saves metadata in the .unionfs directory within each layer. I
think that's why it doesn't have the copy-up problem I experienced.