mkWindowsApp: add inputHashMethod
Currently, mkWindowsApp uses a strict method based on the Nix store path
to dependencies to generate hashes for each of the WINEPREFIX layers.
This provides the maximum reproduceability, with many layer rebuilds
being the cost.

This commit adds an alternative method based on the Wine and package
version. This method is more relaxed and suitable for packages you don't
want rebuilt often due to their size; Games are an excellent example.