mkWindowsApp - use mv instead of rm in file mapping
To symlink an external directory into the WINEPREFIX,
mkWindowsApp was would first remove the destination file/directory;
Otherwise the symlink can't be created since the file/directory
already exists. That worked just fine, but to delete a directory
all of its contents must also  be deleted. This means that on the
filesystem level, deleting a directory requires metadata and data
updates.

This change renames the file/directory instead. It generates a
random postfix using uuidgen when renaming. Renaming a file/directory
only requires a metadata update on that file/directory. No data
updates are required. This is a more efficient approach.