sierrachart: update user's ACS_Source from tar archive
Prior to this change, the user's ACS_Source was being updated by copying
files from the Nix store. This was causing the ACS_Source directory to
loose the write bit, so attempts to add new files would fail.

This change updates the process to extract the ACS_Source from a tar
archive, which addresses this problem.

This change also makes it so that the ACSIL examples are also updated.