mkWindowsApp: add ability to disable the Mono prompt
The smaller Wine packages, such as wine64Packages.base,
don't come with wine-mono. This causes Wine to produce
a warning and prompt during the initialization of the
WINEPREFIX.

This change adds a parameter, enableMonoBootPrompt,
which can be used to disable that prompt. The default behavior
is to do nothing and allow the prompt.

To make this work, the DLL overrides are now part of the
Windows layer hash.